requirements.txt
asgiref==3.3.1
Django==3.1.3
django-bootstrap-form==3.4
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
